:sparkles: feat: Add API route for retrieving categories
This commit adds a new API route for retrieving categories. The `GET` method is implemented to fetch the categories data from the server using the `api.get` function. If successful, the response data is returned as a JSON response. If there is an error, a JSON response with a status code of 500 and a message of "Erro interno do servidor" is returned.